Idk but I added some tests.

+++ SimpleJWT/SimpleJWT/Providers/RSAKeyProvider.cs
@@ -17,7 +17,9 @@ namespace SimpleJWT.Providers
         public RSAKeyProvider()
         {
             var folder = AppDomain.CurrentDomain.BaseDirectory + @"RsaKeys";
-            Directory.Delete(folder, true);
+            if(Directory.Exists(folder))
+                Directory.Delete(folder, true);
+
             Directory.CreateDirectory(folder);
 
             PublicRsaKeyPath = folder + @"\RsaPublicKey.txt";
